Build: Add native autotools build system support
This adds support for building with autotools, both independently and as a git subtree in a parent project that also uses autotools.

No dependency on CMake is required here, and support for cross-compiling is present.

The Relic sub-library is configured and built as part of the process, pursuant to the configuration options defined in the CMake build.

The automatic changes to `contrib/relic/include/relic-conf.h.in` should be ignored and not committed back to the repository.